+++ common/src/main/java/net/bettercombat/logic/Pose.java
@@ -2,4 +2,4 @@ package net.bettercombat.logic;
 
 import net.minecraft.util.Identifier;
 
-public record Pose(Identifier base, Identifier offHand) {  }
+public record Pose(String base, String offHand) {  }